+++ app/features/feedback/shared/strategies/drawing/column/reducer.js
