+++ app/features/feedback/shared/strategies/drawing/validations.js
