--- app/features/feedback/shared/strategies/radial/validations.js
@@ -1,13 +1,0 @@
-function validateDefaultTolerance(formState) {
-  if (!formState.defaultTolerance) {
-    return false;
-  } else {
-    const re = /^\d+(?:\.\d+)?$/;
-    return typeof formState.defaultTolerance === 'string' &&
-      re.test(formState.defaultTolerance);
-  }
-}
-
-export default [
-  validateDefaultTolerance
-];
